Had to fix some reference errors, reainstall bootstrap, moved ddashboard component from containers to components
