+++ front-end/src/App.js
@@ -4,10 +4,9 @@ import './App.css';
 
 import SignUp from './components/Signup/SignUp';
 import LogIn from './components/LogIn/LogIn.js';
-
-import Assignments from './components/Assignments/Assignments';
-import EnsureLoggedInContainer from './components/EnsureLoggedIn/EnsureLoggedInContainer';
 import LandingPage from './components/landingpage/landingpage';
+import TeacherDashboard from './Containers/TeacherDashboard';
+
 class App extends Component {
 	render() {
 		return (
