+++ front-end/src/Containers/TeacherDashboard.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 // import { Route } from 'react-router-dom';
 import SideMenu from '../components/SideMenu/SideMenu';
 import AssignmentList from '../components/Assignments/AssignmentList';
-import Dashboard from './Dashboard';
+import Dashboard from '../components/Dashboard/Dashboard';
 import AddAssignmentForm from '../components/Assignments/AddAssignment/AddAssignmentForm';
 
 import { Grid, Button } from 'material-ui';
